stopwatch_user.c
File to utilize the stopwatch driver with the hardware switches and push down button driver. 

/* Pressing KEY0 should toggle the stopwatch between the run and pause states. 
Pressing KEY1 to KEY3 should set the time according to the values of the SW slider 
switches. Set the hundredths (DD) if KEY1 is pressed, the seconds (SS) for KEY2, 
and the minutes (MM) for KEY3.*/